+++ src/codemagic_cli_tools/apple/__init__.py
@@ -1,5 +1,5 @@
 from .app_store_connect_api import AppStoreConnectApiClient
 from .app_store_connect_api import AppStoreConnectApiError
-from .operations import AppOrdering
-from .operations import BundleIdOrdering
-from .operations import CertificateOrdering
+from .app_store_connect_operations import AppOrdering
+from .app_store_connect_operations import BundleIdOrdering
+from .app_store_connect_operations import CertificateOrdering
