+++ src/codemagic_cli_tools/apple/app_store_connect_operations/bundle_id_capabilities_operations.py
